Rename PostgreSql shell feature classes for consistency
