+++ src/modules/Elsa.Persistence.EFCore.PostgreSql/ShellFeatures/Alterations/PostgreSqlAlterationsPersistenceFeature.cs
@@ -1,22 +1,22 @@
 using System.Reflection;
 using CShells.Features;
 using Elsa.Persistence.EFCore.Extensions;
-using Elsa.Persistence.EFCore.Modules.Tenants;
+using Elsa.Persistence.EFCore.Modules.Alterations;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore;
 
-namespace Elsa.Persistence.EFCore.PostgreSql.ShellFeatures.Tenants;
+namespace Elsa.Persistence.EFCore.PostgreSql.ShellFeatures.Alterations;
 
 /// <summary>
-/// Configures the tenants feature to use PostgreSql persistence.
+/// Configures the alterations feature to use PostgreSql persistence.
 /// </summary>
 [ShellFeature(
-    DisplayName = "PostgreSql Tenant Persistence",
-    Description = "Provides PostgreSql persistence for tenant management",
-    DependsOn = ["TenantManagement"])]
+    DisplayName = "PostgreSql Alterations Persistence",
+    Description = "Provides PostgreSql persistence for workflow alterations",
+    DependsOn = ["Alterations"])]
 [UsedImplicitly]
-public class PostgreSqlTenantPersistenceShellFeature
-    : EFCoreTenantManagementShellFeatureBase
+public class PostgreSqlAlterationsPersistenceFeature
+    : EFCoreAlterationsPersistenceShellFeatureBase
 {
     /// <inheritdoc />
     protected override void ConfigureProvider(DbContextOptionsBuilder builder, Assembly migrationsAssembly, string connectionString, ElsaDbContextOptions? options)
