+++ src/modules/Elsa.Persistence.EFCore.PostgreSql/ShellFeatures/Labels/PostgreSqlLabelPersistenceFeature.cs
@@ -1,22 +1,22 @@
 using System.Reflection;
 using CShells.Features;
 using Elsa.Persistence.EFCore.Extensions;
-using Elsa.Persistence.EFCore.Modules.Runtime;
+using Elsa.Persistence.EFCore.Modules.Labels;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore;
 
-namespace Elsa.Persistence.EFCore.PostgreSql.ShellFeatures.Runtime;
+namespace Elsa.Persistence.EFCore.PostgreSql.ShellFeatures.Labels;
 
 /// <summary>
-/// Configures the runtime feature to use PostgreSql persistence.
+/// Configures the labels feature to use PostgreSql persistence.
 /// </summary>
 [ShellFeature(
-    DisplayName = "PostgreSql Workflow Runtime Persistence",
-    Description = "Provides PostgreSql persistence for workflow runtime",
-    DependsOn = ["WorkflowRuntime"])]
+    DisplayName = "PostgreSql Label Persistence",
+    Description = "Provides PostgreSql persistence for label management",
+    DependsOn = ["Labels"])]
 [UsedImplicitly]
-public class PostgreSqlWorkflowRuntimePersistenceShellFeature
-    : EFCoreWorkflowRuntimePersistenceShellFeatureBase
+public class PostgreSqlLabelPersistenceFeature
+    : EFCoreLabelPersistenceShellFeatureBase
 {
     /// <inheritdoc />
     protected override void ConfigureProvider(DbContextOptionsBuilder builder, Assembly migrationsAssembly, string connectionString, ElsaDbContextOptions? options)
